using socket.erl functions for gen_smtp_server
Adding a 'protocol' option to server init that allows for either
tcp or ssl to be used as default starting connection for the server.
This allows ssl v2 (and other) clients to connect without requiring
more advanced TLS support.

This also allows for smoothly starting multiple servers on multiple
ports with different protocols.