gen_smtp_server now supports multiple sockets
The only difference for existing installations is that the given init
options must be wrapped in a list.

To start a server on multiple ports with a choice of protocols:

[ [{port,25},{protocol,tcp},{domain,"myserver.com"},{address,{0,0,0,0}}],
  [{port,465},{protocol,ssl},{domain,"secure.myserver.com"},{address,{0.0.0.0}}]
]

If I knew how to detect whether the connecting client was hoping for an ssl
connection or tcp immediately we would be able to support both ssl and tcp on
the same socket.