FLAG DAY: change callback module init/3 to init/4 THIS BREAKS STUFF!
This is to allow parameters to be passed to the callback, so you can
have a callback module with configurable behaviour.

Also, terminate and code_change are added to the example callback, as
well as optional relaying behaviour (thanks to Wes James). Also,
the authentication example module has been rolled into the example and,
again, is configurable.