Fix SSL socket upgrade
After a STARTTLS, the socket suddenly started to receive list instead
of binary data, on Erlang 23.2.7.

Still need to figure out exactly why this happens, at least this patch
prevents it. Fixes #250

Note that the receiving side has the same issue; this needs to be
fixed in the ranch library.