gen_smtp_client: stop filtering gen_tcp and ssl options, fixes #271
A lot of new useful options were added to gen_tcp and ssl in the recent
OTP releases, but `smtp_socket` module (that backs gen_smtp_client) used
to only allow a smal set of options and discarded all the not defined in
the whitelist.
We think that it's better to let programmer decide which options they
want to pass - it should e programmer's responsibility if they provide
invalid options.